Data Generation done

+++ Lab2/src/searching_Algos.h
@@ -18,8 +18,8 @@
 struct compare{
     bool operator()(const TreeNode*l,const TreeNode*r)
     {
-        int weight_l=(1+l->weight)*(l->distance_parent);
-        int weight_r=(1+r->weight)*(r->distance_parent);
+        double weight_l=(1+l->weight)*(l->distance_parent);
+        double weight_r=(1+r->weight)*(r->distance_parent);
        if(weight_l>weight_r)
            return true;
        return false;
